--- Compose-Flight-Search-App/app/src/androidTest/kotlin/com/example/flightsearchapp/testing/model/database/AirportEntitiesTestData.kt
@@ -1,14 +1,0 @@
-package com.example.flightsearchapp.testing.model.database
-
-import com.example.flightsearchapp.data.database.AirportEntity
-import com.example.flightsearchapp.testing.model.database.entityextension.testAirportEntity
-
-/**
- * Test data for List<[AirportEntity]>
- */
-val airportEntitiesTestData: List<AirportEntity> = listOf(
-    testAirportEntity(1, "ABC"),
-    testAirportEntity(2, "CDE"),
-    testAirportEntity(3, "EFG"),
-    testAirportEntity(4, "GHI"),
-)
