--- Compose-Flight-Search-App/app/src/androidTest/kotlin/com/example/flightsearchapp/testing/model/database/AirportFtsEntitiesTestData.kt
@@ -1,10 +1,0 @@
-package com.example.flightsearchapp.testing.model.database
-
-import com.example.flightsearchapp.data.database.AirportFtsEntity
-import com.example.flightsearchapp.data.database.asFtsEntity
-import com.example.flightsearchapp.testing.model.database.entityextension.testAirportEntity
-
-val airportFtsEntitiesTestData: List<AirportFtsEntity> = listOf(
-    testAirportEntity(1, "HAM").asFtsEntity(),
-    testAirportEntity(2, "CDE").asFtsEntity(),
-)
