--- Compose-Flight-Search-App/app/src/test/kotlin/com/example/flightsearchapp/testing/model/database/AirportFtsEntitiesTestData.kt
@@ -1,9 +1,0 @@
-package com.example.flightsearchapp.testing.model.database
-
-import com.example.flightsearchapp.data.database.AirportFtsEntity
-import com.example.flightsearchapp.data.database.asFtsEntity
-
-val airportFtsEntitiesTestData: List<AirportFtsEntity> = listOf(
-    testAirportEntity(1, "HAM").asFtsEntity(),
-    testAirportEntity(2, "CDE").asFtsEntity(),
-)
