--- Compose-Flight-Search-App/app/src/test/kotlin/com/example/flightsearchapp/testing/model/database/FavoriteEntitiesTestData.kt
@@ -1,15 +1,0 @@
-package com.example.flightsearchapp.testing.model.database
-
-import com.example.flightsearchapp.data.database.FavoriteEntity
-
-val favoriteEntitiesTestData: List<FavoriteEntity> = listOf(
-    testFavoriteEntity(1, "ABC", "CDE"),
-    testFavoriteEntity(2, "EFG", "GHI"),
-)
-
-private fun testFavoriteEntity(id: Long, departureCode: String, destinationCode: String) =
-    FavoriteEntity(
-        id = id,
-        departureCode = departureCode,
-        destinationCode = destinationCode,
-    )
