+++ Compose-Flight-Search-App/core/database/src/test/kotlin/com/example/flightsearchapp/core/database/AirportDaoTest.kt
@@ -1,13 +1,9 @@
-package com.example.flightsearchapp.database
+package com.example.flightsearchapp.core.database
 
 import androidx.room.Room
-import com.example.flightsearchapp.data.database.AirportDao
-import com.example.flightsearchapp.data.database.AppDatabase
-import com.example.flightsearchapp.testing.model.database.airportEntitiesTestData
-import kotlin.test.assertEquals
-import kotlin.test.assertNotNull
-import kotlin.test.assertNull
-import kotlin.test.assertTrue
+import com.example.flightsearchapp.core.database.dao.AirportDao
+import com.example.flightsearchapp.core.database.dao.AppDatabase
+import com.example.flightsearchapp.core.testing.database.airportEntitiesTestData
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.test.runTest
@@ -17,6 +13,10 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
+import kotlin.test.assertEquals
+import kotlin.test.assertNotNull
+import kotlin.test.assertNull
+import kotlin.test.assertTrue
 
 /**
  * Instrumented test for [AirportDao]
