+++ Compose-Flight-Search-App/core/testing/src/main/java/com/example/flightsearchapp/core/testing/database/AirportFtsEntitiesTestData.kt
@@ -1,0 +1,9 @@
+package com.example.flightsearchapp.core.testing.database
+
+import com.example.flightsearchapp.core.data.model.asFtsEntity
+import com.example.flightsearchapp.core.database.model.AirportFtsEntity
+
+val airportFtsEntitiesTestData: List<AirportFtsEntity> = listOf(
+    testAirportEntity(1, "HAM").asFtsEntity(),
+    testAirportEntity(2, "CDE").asFtsEntity(),
+)
