+++ Compose-Flight-Search-App/core/testing/src/main/java/com/example/flightsearchapp/core/testing/database/entityextension/EntityExtensions.kt
@@ -1,0 +1,11 @@
+package com.example.flightsearchapp.core.testing.database.entityextension
+
+import com.example.flightsearchapp.core.database.model.AirportEntity
+
+fun testAirportEntity(id: Long, code: String) =
+    AirportEntity(
+        id = id,
+        name = "name$id",
+        iataCode = code,
+        passengers = id,
+    )
